🐛 move click to set to not conflict with player overview
