+++ frontend/app/src/components/views/competition/knockoutSystem/ViewKnockoutSystem.vue
@@ -15,7 +15,7 @@
 					v-if="match"
 					:match="match"
 					:mode="props.mode"
-					@click="showPopUp(match)"
+					:numberSets="props.numberSets"
 				/>
 			</template>
 			<template #additional="{ match }">
@@ -23,11 +23,6 @@
 			</template>
 		</ViewKnockoutTree>
 	</template>
-	<UpdatePointsDialog
-		v-if="canEdit"
-		ref="dialog"
-		:number-sets="props.numberSets"
-	/>
 </template>
 
 <script lang="ts" setup>
@@ -39,10 +34,6 @@ import ViewMatch from "@/components/views/competition/knockoutSystem/ViewMatch.v
 import ViewMatchDate from "@/components/views/competition/knockoutSystem/ViewMatchDate.vue"
 import { knockoutTitle } from "@/components/views/competition/knockoutSystem/KnockoutTitleGenerator"
 import { useI18n } from "vue-i18n"
-import { inject, ref } from "vue"
-import { Match } from "@/interfaces/match"
-import UpdatePointsDialog from "@/components/items/UpdatePointsDialog.vue"
-import { getCanEdit } from "@/backend/security"
 
 const { t } = useI18n({ inheritLocale: true })
 
@@ -52,17 +43,8 @@ const props = defineProps<{
 }>()
 
 const route = useRoute()
-const isLoggedIn = inject("loggedIn", ref(false))
-const { data: canEdit } = getCanEdit(<string>route.params.tourId, isLoggedIn)
 
 const { data: knockoutSystem } = getKnockout(route)
-const dialog = ref()
-
-const showPopUp = function (match: Match) {
-	if (canEdit.value) {
-		dialog.value.showPopUp(match)
-	}
-}
 </script>
 
 <style scoped>
